I've renamed the `agents` directory to `crypto_agents` to resolve an import conflict. This prevents a name collision with a common Python library that is also named `agents`.
This conflict was causing a `ValueError: numpy.dtype size changed` error on systems where the other `agents` library (often used for reinforcement learning) is installed. The Python import system was incorrectly loading the installed library instead of the local project files.

I also updated the import statements in `trader.py` and `tester.py` to reflect the new directory name.
